Move `get_orig` fixture to a conftest.py
Allows for sharing amongst test sub-mods. Use `Originator.waitwhile`
where appropriate.
