+++ tests/apps/test_originator.py
@@ -9,43 +9,9 @@
     has been assigned to the switchy dialplan.
 '''
 from __future__ import division
-import pytest
 import time
 import math
 from switchy.apps import dtmf, players
-from switchy import get_originator
-
-
-@pytest.yield_fixture
-def get_orig(request, fsip):
-    '''Deliver an `Originator` app which drives a single
-    FreeSWITCH slave process.
-    '''
-    origs = []
-
-    def factory(userpart, port=5080, limit=1, rate=1, offer=1, **kwargs):
-        orig = get_originator(
-            fsip,
-            limit=limit,
-            rate=rate,
-            max_offered=offer,
-            **kwargs
-        )
-
-        # each slave profile should call originate calls to itself
-        # to avoid dependency on another server
-        orig.pool.evals(
-            ("""client.set_orig_cmd('{}@{}:{}'.format(
-             userpart, client.server, port))"""),
-            userpart=userpart,
-            port=port,
-        )
-        origs.append(orig)
-        return orig
-
-    yield factory
-    for orig in origs:
-        orig.shutdown()
 
 
 def test_rep_fields(get_orig):
@@ -132,8 +98,6 @@ def test_convo_sim(get_orig):
     assert orig.pool.count_calls() == orig.limit
 
     # wait for all calls to end
-    while not orig.stopped() or orig.pool.count_calls():
-        time.sleep(1)
-
+    orig.waitwhile()
     # ensure number of calls recorded matches the rec period
     assert float(len(recs)) == math.floor((stop - start) / playrec.rec_period)
